:recycle: [open-formulieren/formio-builder#271] Move utility code into form designer package
