+++ src/openforms/js/components/admin/form_design/FormBuilder.js
@@ -3,35 +3,37 @@ import {useContext, useState} from 'react';
 import {useGlobalState} from 'state-pool';
 
 import {FormContext} from 'components/admin/form_design/Context';
-import {getUniqueKey} from 'components/admin/form_design/utils';
-import {getMapOverlayTileLayers} from 'components/formio_builder/mapLayers';
 import {
+  getAvailableAuthPlugins,
   getPrefillAttributes,
   getPrefillPlugins,
   getRegistrationAttributes,
   getValidatorPlugins,
-} from 'components/formio_builder/plugins';
+} from 'components/admin/form_design/plugins';
 import {
   getReferenceListsTableItems,
   getReferenceListsTables,
   getServices,
-} from 'components/formio_builder/referenceLists';
+} from 'components/admin/form_design/reference-lists';
+import {getUniqueKey} from 'components/admin/form_design/utils';
 import {flattenComponents} from 'components/utils';
-import {get} from 'utils/fetch';
 import jsonScriptToVar from 'utils/json-script';
 import {currentTheme} from 'utils/theme';
-
-import {AUTH_PLUGINS_ENDPOINT} from './constants';
 
 const LANGUAGES = jsonScriptToVar('languages', {default: []}).map(([langCode]) => langCode);
 const FILE_TYPES = jsonScriptToVar('config-UPLOAD_FILETYPES', {default: []});
 const MAX_FILE_UPLOAD_SIZE = jsonScriptToVar('setting-MAX_FILE_UPLOAD_SIZE', {default: 'unknown'});
 const RICH_TEXT_COLORS = jsonScriptToVar('config-RICH_TEXT_COLORS', {default: []});
 const MAP_TILE_LAYERS = jsonScriptToVar('config-MAP_TILE_LAYERS', {default: []});
+const MAP_WMS_LAYERS = jsonScriptToVar('config-MAP_WMS_LAYERS', {default: []});
+const MAP_WFS_LAYERS = [];
 
-const getAvailableAuthPlugins = async () => {
-  const response = await get(AUTH_PLUGINS_ENDPOINT);
-  return response.data;
+const getMapOverlayTileLayers = async () => {
+  const layers = [
+    ...MAP_WMS_LAYERS.map(layer => ({...layer, type: 'wms'})),
+    ...MAP_WFS_LAYERS.map(layer => ({...layer, type: 'wfs'})),
+  ];
+  return layers.sort((layerA, layerB) => layerA.name.localeCompare(layerB.name));
 };
 
 export const getInitialUsedComponentKeys = components => {
